- Separated ERC read only functions tests in a single class and added balanceOfEmptySpender tests
